+++ lib/controller/model/scenekit_widget_model.dart
@@ -3,15 +3,17 @@ import 'package:flutter/material.dart';
 class ScenekitWidgetModel {
   final double lat;
   final double long;
-  final String name;
   final String? assetName;
   final int? hexColor;
   final VoidCallback? onWidgetTap;
+  String? name;
+
+  set setWidgetName(String widgetName) => name = widgetName;
 
   ScenekitWidgetModel({
     required this.lat,
     required this.long,
-    required this.name,
+    this.name,
     this.assetName,
     this.hexColor,
     this.onWidgetTap,